+++ tests/testthat/test-resample_draws.R
@@ -55,9 +55,11 @@ test_that("Resampling algorithms return the correct result in expectation", {
 
 test_that("resample_draws uses stored weights when available", {
   x <- example_draws()
-  expect_error(resample_draws(x),
-    "No weights are provided and none can be found within the draws object"
-  )
+
+  x_rs <- resample_draws(x)
+  expect_true(is_draws_array(x_rs))
+  expect_equal(ndraws(x_rs), ndraws(x))
+
   w <- runif(ndraws(x), 0, 10)
   x <- weight_draws(x, w)
   x_rs <- resample_draws(x)
@@ -65,4 +67,20 @@ test_that("resample_draws uses stored weights when available", {
   expect_equal(ndraws(x_rs), ndraws(x))
   # .log_weight variable has been dropped
   expect_true(!".log_weight" %in% variables(x_rs, reserved = TRUE))
+})
+
+test_that("resample_draws works on rvars", {
+  set.seed(1234)
+  x <- rvar(1:10000)
+  w <- 1:10000 / 777
+
+  x_rs <- resample_draws(x, w, method = "stratified")
+  mean_rs <- mean(x_rs)
+  expect_true(mean_rs > 6660 && mean_rs < 6670)
+  expect_true(is_rvar(x_rs))
+
+  # without weights
+  x_rs <- resample_draws(x, method = "stratified")
+  mean_rs <- mean(x_rs)
+  expect_true(mean_rs > 4990 && mean_rs < 5010)
 })